Create BackgroundbuttonWidget and restore ButtonWidget
