+++ demo/config/targets/INST/screens/commanding.txt
@@ -14,6 +14,8 @@ VERTICAL
       NAMED_WIDGET COLLECT_TYPE COMBOBOX NORMAL SPECIAL
       LABEL "  Duration: "
       NAMED_WIDGET DURATION TEXTFIELD 12 "10.0"
+      # Button widgets execute their code directly in the GUI thread
+      # This allows us to directly ask for the text on the named widget
       BUTTON 'Start Collect' 'target_name = get_target_name("INST"); cmd("#{target_name} COLLECT with TYPE #{get_named_widget("COLLECT_TYPE").text}, DURATION #{get_named_widget("DURATION").text.to_f}")'
     END
   END
@@ -31,7 +33,10 @@ VERTICAL
   VERTICALBOX
     SECTIONHEADER "Run Script:"
     NAMED_WIDGET SCRIPTNAME COMBOBOX <%= Dir["#{Cosmos::System.paths['PROCEDURES'][0]}/*"].map {|x| File.basename(x) }.join(' ') %>
-    BUTTON 'Run Script' 'Cosmos.run_process("#{Cosmos::USERPATH}/tools/ScriptRunner -r #{get_named_widget("SCRIPTNAME").text}")'
+    # Here as an example we're using the BACKGROUNDBUTTON which runs all the code in the background.
+    # Note we didn't need to use this here but it's an example of wrapping calls to the GUI in Qt.execute_in_main_thread.
+    # You need to use this button if your code blocks or takes a long time to execute so you don't hang the screen.
+    BACKGROUNDBUTTON 'Run Script' 'Cosmos.run_process("#{Cosmos::USERPATH}/tools/ScriptRunner -r #{Qt.execute_in_main_thread { get_named_widget("SCRIPTNAME").text }}")'
   END
   SETTING BACKCOLOR 163 185 163
 END
